% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeCompALK.r
\name{ageLenMultinomFun}
\alias{ageLenMultinomFun}
\title{multinom' function applied to Tab object}
\usage{
ageLenMultinomFun(object, timeStrata = NULL, spaceStrata = NULL,
  techStrata = NULL, elmts = list(tp = "all", sp = "all", tc = "all"),
  age.plus = -1, ...)
}
\description{
multinom' function applied to Tab object
}
