\name{cod2004ce}
\alias{cod2004ce}
%\alias{had2004ce}
%\alias{whi2004ce}
%\alias{sai2004ce}
%\alias{had1999ce}
\docType{data}
\title{FRS commercial effort data}
\description{
FRS commercial effort data in the COST data exchange format.
%\cr Currently consists of cod, haddock, whiting and saithe for 2004, and haddock for 1999. 
}
\usage{
data(cod2004ce)
%data(had2004ce)
%data(whi2004ce)
%data(sai2004ce)
%data(had1999ce)
}
\format{
Formal class 'ceData' [package "COSTcore"] objects with 2 slots:
\cr@desc: description
\cr@ce: data frame of 18 variables  
\cr see ceData for details of the variables  
}
\details{
%#  Test data set only. Not to be used or cited without prior permission. 
Commercial effort data from vessels landing the names species into Scottish ports, the data are 
aggregated by month, ICES statistical rectangle, ICES area and gear type.  
Compiled from fisheries information network (FIN) data.  
}
\section{Warning }{These are test data sets only and should not to be used or cited without prior permission.}

\source{
  FRS Marine Laboratory, Aberdeen, Scotland. 
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(cod2004ce)
%# maybe str(cod2004ce) ; plot(cod2004ce) ...
}
\keyword{datasets}
