\name{demersal.sampling.lines}
\alias{demersal.sampling.lines}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Adds FRS demersal sampling area boundaries to a map}
\description{
Adds FRS demersal sampling area boundaries to a map. 
}
\usage{
demersal.sampling.lines(doarea,nos=FALSE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{doarea}{An optional numeric vector of the sampling areas boundaries to draw, the default is all.}
  \item{nos}{Logical, adds area numbers if TRUE, default is FALSE.}
  \item{...}{Other arguments to lines and text e.g. ldw, lty.}
}
%\value{
 % ~Describe the value returned
  %If it is a LIST, use
  %\item{comp1 }{Description of 'comp1'}
  %\item{comp2 }{Description of 'comp2'}
%}

\author{Alastair Pout\email{a.pout@marlab.ac.uk}}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }

%\examples{
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{aplot}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
