% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketSeaSampGraphs_Explore.r
\name{landisVolumeFun}
\alias{landisVolumeFun}
\title{ind2', the valid sampling indicator, is built}
\usage{
landisVolumeFun(object, species, fraction, timeStrata, spaceStrata, techStrata,
  sampPar = TRUE, ...)
}
\description{
ind2', the valid sampling indicator, is built
}
