% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeCompALK.r
\name{plotAlMulti}
\alias{plotAlMulti}
\title{datatfit' is the same table as datat, but with fitted values (from multinomial modelisation)}
\usage{
plotAlMulti(x, grps = NULL, show.legend = "right", ...)
}
\description{
datatfit' is the same table as datat, but with fitted values (from multinomial modelisation)
}
