\name{sole3.ce}
\alias{sole3.ce}
\docType{data}
\title{Sole's commercial effort datasets used for COST testing and developing}
\description{
Sole's commercial data with efforts statistics records.
}
\usage{data(sole3.ce)}
\format{sole3.ce ...Formal class 'ceData' [package "COSTcore"] with 2 slots: \code{desc} and \code{ce}.
}

\section{WARNING}{
The data provided is for testing purposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole3.ce)
}
\keyword{datasets}
