\name{sole3.cl}
\alias{sole3.cl}
\docType{data}
\title{Sole's commercial landings datasets used for COST testing and developing}
\description{
Sole's commercial data with landings statistics records.
}
\usage{data(sole3.cl)}
\format{sole3.cl ...Formal class 'clData' [package "COSTcore"] with 2 slots: \code{desc} and \code{cl}.
}

\section{WARNING}{
The data provided is for testing purposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole3.cl)
}
\keyword{datasets}
