\name{sole3.cs}
\alias{sole3.cs}
\docType{data}
\title{Sole's commercial sampling datasets used for COST testing and developing}
\description{
Sole's commercial sampling data.).
}
\usage{data(sole3.cs)}
\format{sole3.cs ...Formal class 'csData' [package "COSTcore"] with 6 slots: \code{desc}, \code{tr}, \code{hh}, \code{sl}, \code{hl} and \code{ca}.
}

\section{WARNING}{
The data provided is for testing purposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole3.cs)
}
\keyword{datasets}
