\name{cod2004ce}
\alias{cod2004ce}
\docType{data}
\title{Cod 2004 commercial effort data}
\description{
  Cod 2004 commercial effort data in the COST data exchange format.
}
\usage{data(cod2004ce)}
\format{
Formal class 'ceData' [package "COSTcore"] with 2 slots:
\cr@desc: description
\cr@ce: data frame of 2505 observations of 15 variables  
\cr see ceData for details of the variables  
 % .. ..$ vslFlgCtry: chr [1:2505] "SCO" "SCO" "SCO" "SCO" ...
 % .. ..$ year      : num [1:2505] 2004 2004 2004 2004 2004 ...
%  .. ..$ quarter   : num [1:2505] 2 2 2 2 2 2 2 2 2 2 ...
  %.. ..$ month     : int [1:2505] 4 4 4 4 4 4 4 4 4 4 ...
 % .. ..$ area      : chr [1:2505] "IVb" "IVa" "VIIa" "IVa" ...
  %.. ..$ rect      : chr [1:2505] "40F4" "48F2" "37E5" "47E9" ...
  %.. ..$ foCatNat  : chr [1:2505] "TBB" "OTB" "SSC" "OTB" ...
  %.. ..$ foCatEu5  : chr [1:2505] "TBB" "OTB" "SSC" "OTB" ...
  %.. ..$ foCatEu6  : chr [1:2505] "TBB" "OTB" "SSC" "OTB" ...
  %.. ..$ trpNum    : num [1:2505] 1 4 3 41 6 1 26 4 59 10 ...
  %.. ..$ foNum     : int [1:2505] NA NA NA NA NA NA NA NA NA NA ...
  %.. ..$ foDur     : int [1:2505] NA NA NA NA NA NA NA NA NA NA ...
  %.. ..$ effKwDays : int [1:2505] NA NA NA NA NA NA NA NA NA NA ...
  %.. ..$ effGtDays : int [1:2505] NA NA NA NA NA NA NA NA NA NA ...
  %.. ..$ daysAtSea : num [1:2505]   3.67  15.65   7.50 180.67  19.90 ...
}
\details{
%#  Test data set only. Not to be used or cited without prior permission. 
Commercial effort data from vessels landing cod into Scottish ports during 2004,  
aggregated by month, ICES statistical rectangle, ICES area and gear type.  
Compiled from fisheries information network (FIN) data.  
}
\section{Warning }{This is a test data set only and should not to be used or cited without prior permission.}

\source{
  FRS Marine Laboratory, Aberdeen, Scotland. 
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(cod2004ce)
%# maybe str(cod2004ce) ; plot(cod2004ce) ...
}
\keyword{datasets}
