\name{cod2004cs}
\alias{cod2004cs}
\docType{data}
\title{Cod 2004 commercial sampling data }
\description{
  Cod 2004 commercial sampling data in the COST data exchange format.
}
\usage{data(cod2004cs)}
\format{
Formal class 'csData' [package "COSTcore"] with 6 slots
\cr@desc: description
\cr@tr: data.frame of 443 observations of  15 variables
\cr@hh: data.frame of 443 observations of  29 variables
\cr@sl: data.frame of 1175 observations of  16 variables
\cr@hl: data.frame of 14913 observations of  16 variables
\cr@ca: data.frame of 13925 observations of  28 variables
\cr see csData for details of the variables   
}
\details{
Commercial sampling data from vessels landing cod into Scottish ports during 2004,  
aggregated by month, ICES statistical rectangle, ICES area and gear type.  
Compiled from FRS sampling programme data. 
}
\section{Warning }{This is a test data set only and should not to be used or cited without prior permission.}

\source{
  FRS Marine Laboratory, Aberdeen. Scotland. 
}



%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(cod2004cs)
%## maybe str(cod2004cs) ; plot(cod2004cs) ...
}
\keyword{datasets}
